webCrawling/pyproject.toml
[project]
name = "webCrawling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "crawl4ai>=0.7.2",
    "dspy>=2.6.27",
    "feedparser>=6.0.11",
    "openai>=1.97.1",
    "playwright>=1.54.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "selenium>=4.34.2",
    "supabase>=2.17.0",
    "webdriver-manager>=4.0.2",
]
